Restructure Flask app
