+++ casino/games/__init__.py
@@ -1,0 +1,2 @@
+from .roullete import ruleta_bp as roullete_blueprint
+from .piedra_papel_tijeras import ppt as ppt_blueprint
