+++ casino/static/js/ruleta.js
