+++ casino/templates/juegos.html
@@ -11,7 +11,7 @@
 
     <button onclick="location.href='/ruleta'">Ruleta</button>
     <button onclick="location.href='/piedra-papel-tijera'">Piedra, Papel o Tijera</button>
-    <button onclick="location.href='/mp-ppt'">ONLINE Piedra, Papel o Tijera</button>
+    <button onclick="location.href='/online-ppt'">ONLINE Piedra, Papel o Tijera</button>
 
 </body>
