+++ casino/templates/online-ppt.html
@@ -1,0 +1,12 @@
+<!-- templates/rock_paper_scissors.html -->
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Piedra, Papel o Tijera ONLINE</title>
+</head>
+<body>
+    <h1> Piedra, Papel o Tijera ONLINE</h1>
+
+</body>
+</html>
