--- templates/mp_ppt.html
@@ -1,18 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Casino Hub</title>
-</head>
-
-<body>
-    <h1>Seleccionar Juego</h1>
-
-
-    <button onclick="location.href='/ruleta'">Ruleta</button>
-    <button onclick="location.href='/piedra-papel-tijera'">Piedra, Papel o Tijera</button>
-    <button onclick="location.href='/mp-ppt'">ONLINE Piedra, Papel o Tijera</button>
-
-</body>
-
-</html>
